Fixing test suite issue with accepting/connecting.
Because of a weird TCP stack behaviour it is possible that
connect returns with a socket before it is properly accepted on the
server side. On top of this if the listening socket is closed between
the connect returning and the accept really happening the connection
disappears on the server side. Like it never happened in the first place.