Ignore `content-length` when TE chunked is present
Some servers return BOTH `transfer-encoding: chunked` as well as
`content-length`, which in error, but not much help for the client.
This fixes the response handling so that it ignores the
`content-length` header if `transfer-encoding: chunked` is present
in the response, which makes the client "work" in the presence of
such a misbehaving server.