Fix header processing to always be case-insensitive
This problem occurred in two ways: first, trailers did not get
lower cased when read; and second, `lhttpc_lib:header_value/3
did not lower-case the requested header name, so that it would
only be found if it was an exact case match.  How no one ever
complained about this behavior is why we cannot have nice things.